streamlit/requirements.txt
numpy
pandas
Pillow
pymupdf
streamlit
